Start Application from separate class, so Javafx runtime components are found correctly by Gradle
